fix(tui): wire status bar context %, streaming state, and token types (#149)
* fix(tui): wire status bar context %, streaming state, and token types (#134)

- Wire StreamEvent::TokenUsage → StatusState::Streaming on TextDelta
  events so the status bar transitions from Thinking to Streaming with
  an approximate token count while the LLM is generating text
- Compute context window usage percentage at TurnComplete using model
  metadata lookup (via genesis_provider::model_metadata)
- Promote token counters from u32 to u64 to prevent overflow on long sessions
- Remove /compact from command popup (was a silent no-op)
- Remove Submission::Compact dead code
- Make model_metadata module public in genesis-provider
- Wrap SkillStore::list_all() in spawn_blocking to avoid blocking async runtime
- Add viewport_width tracking to App for transcript overlay sizing
- Use actual terminal width for transcript overlay instead of hardcoded 80
- Pass real model name to StatusBarWidget instead of "default"

Closes #134
Closes #135

* fix(tui): use latest turn's input_tokens for context % and strip model prefix

Address review feedback:

1. Context % now uses the latest turn's input_tokens (which already
   includes the full conversation context) instead of the cumulative
   session sum. Each LLM call re-sends the full history, so summing
   across turns would double-count and over-report context usage.

2. Strip provider prefix from model ID before metadata lookup
   (e.g. "anthropic/claude-sonnet-4-6" → "claude-sonnet-4-6") so
   OpenRouter-style configs correctly resolve context window sizes.

+++ crates/genesis-tui/src/widgets/status_bar.rs
@@ -77,9 +77,9 @@ pub struct StatusBarWidget {
     /// Git branch name (or cwd fallback). `None` until first populated.
     right_info: Option<String>,
     /// Cumulative input tokens this session.
-    pub tokens_in: u32,
+    pub tokens_in: u64,
     /// Cumulative output tokens this session.
-    pub tokens_out: u32,
+    pub tokens_out: u64,
     /// Elapsed time since the current turn started.
     pub turn_elapsed: Option<Duration>,
 }
@@ -435,7 +435,7 @@ fn write_spans(spans: &[Span<'_>], start_x: u16, row: u16, bound_x: u16, buf: &m
 }
 
 /// Format token counts compactly: 1234 → "1.2k", 12345 → "12k".
-fn format_tokens(n: u32) -> String {
+fn format_tokens(n: u64) -> String {
     if n < 1000 {
         n.to_string()
     } else if n < 10_000 {
